do.not.cat.size_t

+++ includes/ft_malloc.h
@@ -6,7 +6,7 @@
 /*   By: cdeniau <cdeniau@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2015/08/05 09:41:22 by cdeniau           #+#    #+#             */
-/*   Updated: 2015/08/19 19:04:20 by cdeniau          ###   ########.fr       */
+/*   Updated: 2015/09/10 19:05:26 by cdeniau          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -29,7 +29,7 @@
 typedef struct		s_header
 {
 	void			*next;
-	int				size;
+	size_t			size;
 	int				flg;
 }					t_header;
 
@@ -37,14 +37,14 @@ typedef struct		s_tiny
 {
 	t_header		*firstblock;
 	void			*next;
-	int				totalsize;
+	size_t			totalsize;
 }					t_tiny;
 
 typedef struct		s_small
 {
 	t_header		*firstblock;
 	void			*next;
-	int				totalsize;
+	size_t			totalsize;
 }					t_small;
 
 typedef struct		s_large
@@ -69,15 +69,15 @@ void				print_mem(void *mem);
 void				*malloc(size_t size);
 void				*realloc(void *ptr, size_t size);
 void				*calloc(size_t count, size_t size);
-void				*set_header(void *firstblock, int size, int first);
+void				*set_header(void *firstblock, size_t size, int first);
 void				*realloc(void *ptr, size_t size);
 void				*malloc(size_t size);
-void				*ft_malloc_tiny(int size);
-void				*ft_malloc_small(int size);
-void				*ft_malloc_large(int size);
-t_tiny				*ft_new_tiny(int size);
-t_small				*ft_new_small(int size);
-t_large				*ft_new_large(int size);
+void				*ft_malloc_tiny(size_t size);
+void				*ft_malloc_small(size_t size);
+void				*ft_malloc_large(size_t size);
+t_tiny				*ft_new_tiny(size_t size);
+t_small				*ft_new_small(size_t size);
+t_large				*ft_new_large(size_t size);
 void				*ft_overninethousand(void);
 t_tiny				*ft_tiny_find(t_tiny *tiny_head);
 t_small				*ft_small_find(t_small *small_head);
